requirements.txt
pandas
pre-commit
psycopg2-binary
pytest
pytest-cov
python-dotenv
sqlalchemy
